+++ src/main/java/com/studdy/springboot/dao/StudentDAO.java
@@ -6,7 +6,7 @@ import com.studdy.springboot.modal.Student;
 
 public interface StudentDAO {
 	List<Student> get();
-	Student get(Long id);
+	Student get(int id);
 	void save(Student student);
-	void delete(Long id);
+	void delete(int id);
 }
